Added custom Exceptions and added them for handlers

+++ Whisky.Collection.Application/Features/MyWhisky/Commands/DeleteMyWhisky/DeleteMyWhiskyCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using Whisky.Collection.Application.Contracts.Persistence;
+using Whisky.Collection.Application.Exceptions;
 
 namespace Whisky.Collection.Application.Features.MyWhisky.Commands.DeleteMyWhisky;
 
@@ -18,6 +19,8 @@ public class DeleteMyWhiskyCommandHandler : IRequestHandler<DeleteMyWhiskyComman
         var myWhiskyToDelete = await _myWhiskyRepository.GetByIdAsync(request.Id);
 
         // Verify that record exists
+        if (myWhiskyToDelete == null)
+            throw new NotFoundException(nameof(MyWhisky), request.Id);
 
         // Remove to database
         await _myWhiskyRepository.DeleteAsync(myWhiskyToDelete);
